+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - fixed a key item softlock in Crash Site (#662)
 - fixed incorrect item and mesh positions in Home Sweet Home when mirrored (#676)
 - fixed uncontrolled SFX in gym/assault course levels not being linked to the correct setting (#684)
+- fixed the scion being difficult to shoot if Lara only has the shotgun (#696)
 - improved the layout of some options in the UI (#694)
 
 ## [V1.8.4](https://github.com/LostArtefacts/TR-Rando/compare/V1.8.3...V1.8.4) - 2024-02-12